refactored BoxDisplay
